+++ CHANGELOG.md
@@ -14,6 +14,14 @@ All notable changes to this project will be documented in this file. This projec
 
 Nothing yet.
 
+## [1.24.1] - 2026-04-12
+
+### Changed
+
+- Skipped AutoPkg recipe type convention checks when type is unknown. (#55)
+- Updated release documentation to reflect new automated workflow.
+- Built error handling into release automation workflow.
+
 ## [1.24.0] - 2026-04-12
 
 ### Added
@@ -468,7 +476,8 @@ Nothing yet.
 
 - Initial release
 
-[Unreleased]: https://github.com/homebysix/pre-commit-macadmin/compare/v1.24.0...HEAD
+[Unreleased]: https://github.com/homebysix/pre-commit-macadmin/compare/v1.24.1...HEAD
+[1.24.1]: https://github.com/homebysix/pre-commit-macadmin/compare/v1.24.0...v1.24.1
 [1.24.0]: https://github.com/homebysix/pre-commit-macadmin/compare/v1.23.0...v1.24.0
 [1.23.0]: https://github.com/homebysix/pre-commit-macadmin/compare/v1.22.0...v1.23.0
 [1.22.0]: https://github.com/homebysix/pre-commit-macadmin/compare/v1.21.0...v1.22.0